requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
openai>=1.0.0
google-generativeai>=0.7.0
anthropic>=0.7.0
sentence-transformers>=2.2.2
torch>=2.0.0
python-dotenv==1.0.0
tiktoken>=0.5.0
langchain-text-splitters>=0.0.1
pytest==7.4.0
requests>=2.31.0
gunicorn>=21.2.0
